requirements.txt
asgiref==3.8.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
cffi==1.17.1
chardet==3.0.4
click==8.1.6
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
cssselect==1.2.0
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
distro-info==1.7+build1
Django==5.1.5
django-admin-interface==0.29.2
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==23.5
django-guardian==2.4.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
duplicity==2.1.4
Faker==35.0.0
fasteners==0.18
filelock==3.16.1
fonttools==4.55.3
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
mysqlclient==2.2.6
netaddr==0.8.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
packaging==24.2
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydyf==0.11.0
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.1.1
pyphen==0.17.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-slugify==8.0.4
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
scour==0.38.2
setuptools==68.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.3
systemd-python==235
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toposort==1.10
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.7
usb-creator==0.3.16
uuid==1.30
virtualenv==20.28.1
wadllib==1.3.6
weasyprint==63.1
webencodings==0.5.1
wheel==0.42.0
whitenoise==6.8.2
xdg==5
xkit==0.0.0
youtube-transcript-api==0.6.3
zopfli==0.2.3.post1
